requirements.txt
astronomer-cosmos==1.2.5